added access control

+++ app/Http/Controllers/PostsController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use Auth;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         $posts = Post::latest()->get();
@@ -48,17 +54,27 @@ class PostsController extends Controller
     {
         $post = Post::findOrFail($id);
 
+        if($post->user_id != Auth::user()->id)
+        {
+            return redirect('/');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
     public function update(Request $request, $id)
     {
+        $post = Post::findOrFail($id);
+
+        if($post->user_id != Auth::user()->id)
+        {
+            return redirect('/');
+        }
+
         $data = $request->validate([
             'title' => 'required|string|min:8',
             'body' => 'required|string|min:8'
         ]);
-
-        $post = Post::findOrFail($id);
 
         $post->title = $data['title'];
         $post->body = $data['body'];
@@ -71,6 +87,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
+
+        if($post->user_id != Auth::user()->id)
+        {
+            return redirect('/');
+        }
 
         $post->delete();
